Released "OAFC_v4"
Successful release of a working version of AFC (AutoFarmCatizen) with additional optimization. This thread of scripts is now called OAFC (OptimizedAutoCatizenFarm) and works quite faster than the original one. This version is fresh and you may encounter bugs while using it. And if you do, be sure to tell the world about them! Both this and main scripts are considered as working so you are to choose which one to use by yourself.

NOTE: For now, OAFC is kind of a different project from AFC as the code differs. Who knows, maybe in the future these can be merged into one, and become the ultimate OAFC script.